infrastructure-resources/requirements.txt
cdktf~=0.0.18
python-decouple